Healthcare data analysis
This project analyzes data from the Hospital Compare website (http://hospitalcompare.hhs.gov) run by the U.S. Department of Health and Human Services. The website's purpose is to provide data and information about the quality of care at over 4,000 Medicare-certified hospitals in the U.S. This dataset covers all major U.S. hospitals. This dataset is used for a variety of purposes, including determining whether hospitals should be fined for not providing high-quality care to patients (see http://goo.gl/jAXFX
for some background on this particular topic).
The Hospital Compare website contains a lot of data and we will only look at a small subset for this project. 

The three files
• outcome-of-care-measures.csv: Contains information about 30-day mortality and readmission rates
for heart attacks, heart failure, and pneumonia for over 4,000 hospitals.
• hospital-data.csv: Contains information about each hospital.
• Hospital_Revised_Flatfiles.pdf: Contains descriptions of the variables in each file

There are three functions created in this project.
best, rankhospital, and rankall
Each function is stored in a separate file.

The best function:
Takes two arguments: an outcome name (outcome) and a hospital ranking (num). The function reads the outcome-of-care-measures.csv file and returns a 2-column data frame containing the hospital in each state that has the ranking specified in num. For example the function call rankall("heart attack", "best") would return a data frame containing the names of the hospitals that are the best in their respective states for 30-day heart attack death rates.
Hospitals that do not have data on a particular outcome are excluded from the set of hospitals when deciding the rankings.

The rankhospital function: 
This function takes three arguments: the 2-character abbreviated name of a state (state), an outcome (outcome), and the ranking of a hospital in that state for that outcome (num). The function reads the outcome-of-care-measures.csv file and returns a character vector with the name of the hospital that has the ranking specified by the num argument. For example, the call rankhospital("MD", "heart failure", 5) would return a character vector containing the name of the hospital with the 5th lowest 30-day death rate for heart failure. The num argument can take values \best", \worst", or an integer indicating the ranking. If the number given by num is larger than the number of hospitals in that state, then the function returns NA. Hospitals that do not have data on a particular outcome are excluded from the set of hospitals when deciding the rankings. Ties are broken with hospital name.

The rankall function: 
This function takes two arguments: an outcome name (outcome) and a hospital ranking (num). The function reads the outcome-of-care-measures.csv file and returns a 2-column data frame containing the hospital in each state that has the ranking specified in num. For example the function call rankall("heart attack", "best") would return a data frame containing the names of the hospitals that are the best in their respective states for 30-day heart attack death rates.